Instruction Memory
Added and test-benched 4KB of instruction memory. Was originally trying to do this with an IP BRAM, but that proved to be more annoying than it was worth.

Added script to generate .mem files for custom programs. The .mem files then get referenced by the imem.v code to load the rom with instructions.