Parameterization + Data Mem
Setup parameterization for data mem, instruction mem, if_stage, and regfile.

Completed an initial completed data memory file. This removes the dual port architecture I was going to do and instead uses a 1R1RW architecture.

In the future, I will use this as a unified memory space for instructions and data, but for now I will remain with the Harvard architecture and separate the two.