ID Stage Contemplations
The ID stage refuses to cooperate and work correctly. The regfile works on its own, but trying to pull values from it on the ID stage is pulling the wrong numbers. I can SEE on the testbench that the signal within the dut to select rs1 and rs2 is correct, but the numbers coming out are not. Which confuses me....