ID Stage reforging
I think one of my issues is with the general format of the ID stage. I think I need to separate the ID stage into just modules, like how I have the IF stage. I need to testbench and verify independently the instruction decoder, and then combine that for the ID stage, I am currently not doing that.