Refactoring ID stage and TB
Changing the ID stage and its testbench to work now and test. Utilizing the same instruction as the decoder, at least at first.

Maybe not the best value for that for the test vectors. Trying to figure out how to efficiently load each reg.

Good end to the week. Need to rest up and get after it again on Monday.